Add files via upload
Additional Code Change so a Custom FrontEnd REACT/VITE frontend may be applied and used even in local mode, The SarahMemoryUIupdate.py allows for automatic updating of the FrontEnd UI, Please note that SarahMemoryGlobals.py Will have to also be updated for this 3rd style UI to be used. Therefore the End-User may use a Traditional GUI, the Classic Java Script WebUI, or have a Custom Built React/Tailwind WebUI when Configuration Flags are set correctly.