Add files via upload
Minor Bugs and Update Patches, Please Note: the patches for This Folder is to have the Web UI react and behave better instead of going off of just the API, There are Many little issues in which I am working on to make the Program work better, The Foundation is pretty solid, The MODS folder '../data/mods/' will all have the version Number as of now '/v800' and in there will be Patches to fix issues I discover on the system. I do expect to Expand on these known issues more and more. the Folder, '../data/addons' will be designed so Drivers for the system to integrate into, for Example I plan on one of the first Drivers will be to integrate the SarahMemory AiOS for Audino IDE to assist in programming boards and circuits, after that I do plan on adding in even more drivers. After the 1st Driver is published It should be fairly easy for anyone to make this AI system intergrate into other systems.